test: Move MockLogger
It is not related to the frontend build
