+++ flow-server/src/test/java/com/vaadin/flow/server/VaadinSessionTest.java
@@ -53,8 +53,8 @@ import com.vaadin.flow.component.ComponentTest.TestComponent;
 import com.vaadin.flow.component.UI;
 import com.vaadin.flow.function.DeploymentConfiguration;
 import com.vaadin.flow.internal.CurrentInstance;
+import com.vaadin.flow.internal.MockLogger;
 import com.vaadin.flow.server.communication.AtmospherePushConnection;
-import com.vaadin.flow.server.frontend.MockLogger;
 import com.vaadin.flow.server.startup.ApplicationConfiguration;
 import com.vaadin.flow.shared.communication.PushMode;
 import com.vaadin.flow.testcategory.SlowTests;
